# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/purchase_company_settings:
    get:
      tags:
        - PurchaseCompanySettings
      summary: List PurchaseCompanySettings
      description: Get a paginated list of PurchaseCompanySettings with optional filtering and sorting
      operationId: listPurchaseCompanySettings
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseCompanySettingList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PurchaseCompanySettings
      summary: Create PurchaseCompanySetting
      description: Create a new PurchaseCompanySetting
      operationId: createPurchaseCompanySetting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePurchaseCompanySettingRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseCompanySetting'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/purchase_orders:
    get:
      tags:
        - PurchaseOrders
      summary: List PurchaseOrders
      description: Get a paginated list of PurchaseOrders with optional filtering and sorting
      operationId: listPurchaseOrders
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrderList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PurchaseOrders
      summary: Create PurchaseOrder
      description: Create a new PurchaseOrder
      operationId: createPurchaseOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePurchaseOrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PurchaseCompanySetting:
      type: object
      required:
        - id
        - company_id
        - double_validation
        - double_validation_amount
        - company_currency
        - send_reminder
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        double_validation:
          type: string
        double_validation_amount:
          type: string
        company_currency:
          type: string
          maxLength: 3
        send_reminder:
          type: boolean
        metadata:
          type: object
...
    CreatePurchaseCompanySettingRequest:
      type: object
      required:
        - company_id
        - double_validation
        - double_validation_amount
        - company_currency
        - send_reminder
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        double_validation:
          type: string
        double_validation_amount:
          type: string
        company_currency:
          type: string
          maxLength: 3
        send_reminder:
          type: boolean
        metadata:
          type: object
...
    PurchaseCompanySettingList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseCompanySetting'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PurchaseOrder:
      type: object
      required:
        - id
        - po_number
        - order_kind
        - company_id
        - supplier_id
        - status
        - receipt_status
        - invoice_status
        - order_date
        - currency
        - currency_rate
        - acknowledged
        - locked
        - subtotal
        - tax_rate
        - tax_amount
        - total
        - metadata
      properties:
        id:
          type: string
          format: uuid
        po_number:
          type: string
          maxLength: 40
        supplier_quotation_id:
          nullable: true
          type: string
          format: uuid
        order_kind:
          type: string
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        supplier_id:
          type: string
          format: uuid
        status:
          type: string
        receipt_status:
          type: string
        invoice_status:
          type: string
        order_date:
          type: string
          format: date
        schedule_date:
          nullable: true
          type: string
          format: date
        currency:
          type: string
          maxLength: 3
        currency_rate:
          type: string
        acknowledged:
          type: boolean
        locked:
          type: boolean
        date_approve:
          nullable: true
          type: string
          format: date
        agreement_id:
          nullable: true
          type: string
          format: uuid
        subtotal:
          type: string
        tax_rate:
          type: string
        tax_amount:
          type: string
        total:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    CreatePurchaseOrderRequest:
      type: object
      required:
        - po_number
        - order_kind
        - company_id
        - supplier_id
        - status
        - receipt_status
        - invoice_status
        - order_date
        - currency
        - currency_rate
        - acknowledged
        - locked
        - subtotal
        - tax_rate
        - tax_amount
        - total
        - metadata
      properties:
        po_number:
          type: string
          maxLength: 40
        supplier_quotation_id:
          nullable: true
          type: string
          format: uuid
        order_kind:
          type: string
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        supplier_id:
          type: string
          format: uuid
        status:
          type: string
        receipt_status:
          type: string
        invoice_status:
          type: string
        order_date:
          type: string
          format: date
        schedule_date:
          nullable: true
          type: string
          format: date
        currency:
          type: string
          maxLength: 3
        currency_rate:
          type: string
        acknowledged:
          type: boolean
        locked:
          type: boolean
        date_approve:
          nullable: true
          type: string
          format: date
        agreement_id:
          nullable: true
          type: string
          format: uuid
        subtotal:
          type: string
        tax_rate:
          type: string
        tax_amount:
          type: string
        total:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    PurchaseOrderList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseOrder'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []